Separate test views and utils
